# src/my_ajax_api/static/openapi.yaml
...
openapi: 3.1.0
info:
  title: Flask AJAX API
  version: "1.0"
  description: 課程講義《AJAX + RESTful API 使用 Flask》的範例 API
servers:
  - url: http://localhost:8000

# 全域套用 Bearer：所有端點都顯示鎖頭。沒加 @login_required 的端點不帶 token 一樣能呼叫。
security:
  - Bearer: []
...
paths:
...
  /api/fruits:
    get:
      tags: [Fruits]
      summary: 列表：搜尋、排序、分頁
      parameters:
        - { name: name, in: query, schema: { type: string }, description: 名稱包含關鍵字（不分大小寫） }
        - { name: sort, in: query, schema: { type: string, enum: [id, name, price, price_desc], default: id } }
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 5, minimum: 1, maximum: 100 } }
      responses:
        "200":
          description: 分頁結果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagedFruits" }
    post:
      tags: [Fruits]
      summary: 新增（名稱 1 到 20 字、價格 1 到 10000）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FruitInput" }
      responses:
        "201":
          description: 已建立，Location 標頭指向新資源
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Fruit" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /api/fruits/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [Fruits]
      summary: 取得單筆
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Fruit" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Fruits]
      summary: 完整更新
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FruitInput" }
      responses:
        "204": { description: 已更新 }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Fruits]
      summary: 刪除
      responses:
        "204": { description: 已刪除 }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/fruits/slow:
    get:
      tags: [Fruits]
      summary: 等待指定秒數後回傳全部，供同步 vs 非同步示範（4-3）
      parameters:
        - { name: seconds, in: query, schema: { type: integer, default: 10, minimum: 1, maximum: 30 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              example: { waitedSeconds: 3, items: [{ id: 1, name: 蘋果, price: 30 }] }
  /api/fruits/upload:
    post:
      tags: [Fruits]
      summary: 上傳圖片（jpg、png、gif、webp，最大 2 MB），存到 static/uploads（6-1）
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                note: { type: string }
      responses:
        "200":
          description: 上傳成功
          content:
            application/json:
              example: { url: /uploads/3f2a....png, size: 12345, note: 水果圖片 }
        "400": { $ref: "#/components/responses/ValidationError" }
        "413": { description: 超過 2 MB }
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 先呼叫 POST /api/auth/login 取得 token，貼到下方欄位（不用加 Bearer 前綴）
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: 呼叫 POST /api/session/login 後瀏覽器會自動保存並帶上，這裡不需要手動填
  responses:
    ValidationError:
      description: 驗證失敗，errors 的 key 是欄位名
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            title: 輸入資料有誤
            status: 400
            errors: { name: [名稱不可為空], price: [價格必須大於 0] }
    NotFound:
      description: 找不到資源
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
          example: { title: 找不到資源, status: 404 }
...
  schemas:
    Problem:
      type: object
      properties:
        title: { type: string }
        status: { type: integer }
        errors:
          type: object
          additionalProperties: { type: array, items: { type: string } }
    Fruit:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        price: { type: number }
    FruitInput:
      type: object
      required: [name, price]
      properties:
        name: { type: string, maxLength: 20, examples: [龍眼] }
        price: { type: number, exclusiveMinimum: 0, maximum: 10000, examples: [60] }
    PagedFruits:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Fruit" } }
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
        totalPages: { type: integer }
...